% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygamma_functions.R
\name{alpha_to_transitions}
\alias{alpha_to_transitions}
\title{Convert alpha matrix to alpha matrix}
\usage{
alpha_to_transitions(alpha_mat, profiles)
}
\arguments{
\item{alpha_mat}{alpha_mat}

\item{profiles}{Profile list}
}
\description{
Add detail
}
