% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygamma_functions.R
\name{beta_gibbs_dist}
\alias{beta_gibbs_dist}
\title{Return mean and variance of beta's Gibbs distribution}
\usage{
beta_gibbs_dist(ystar, A, z, beta_mat, priorsd_beta)
}
\arguments{
\item{ystar}{sampled auxiliary values}

\item{A}{Profile design matrix}

\item{z}{z values}

\item{beta_mat}{Old beta_mat}

\item{priorsd_beta}{Variance priors}
}
\description{
Add detail
}
