% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygamma_functions.R
\name{build_Xlist_01}
\alias{build_Xlist_01}
\title{Convert alpha matrix to alpha matrix}
\usage{
build_Xlist_01(Xbase, Nskill, Ntime)
}
\arguments{
\item{Xbase}{Design matrix for (0,1) for each skill}

\item{Nskill}{Nskill}

\item{Ntime}{Ntime}
}
\description{
Add detail
}
