% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_variational.R
\name{cdm_ELBO_logistic}
\alias{cdm_ELBO_logistic}
\title{Generate initial values from longitudinal cdm model}
\usage{
cdm_ELBO_logistic(vb_vec, Xdata, paraminfo, Q, nsamples = 20)
}
\arguments{
\item{Xdata}{Data list}

\item{Q}{Q-matrix}

\item{variational_dist}{Variational distributions, including means and vars}

\item{init_vals}{Initial values}
}
\value{
Returns value of ELBO function at variational dists
}
\description{
Generate initial values from longitudinal cdm model
}
