% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funs.R
\name{convert_params_to_vector}
\alias{convert_params_to_vector}
\title{Convert parameter list to vector}
\usage{
convert_params_to_vector(params)
}
\arguments{
\item{params}{params}
}
\description{
Convert parameter list to vector
}
