% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funs.R
\name{convert_vector_to_params}
\alias{convert_vector_to_params}
\title{Convert vector to parameter list}
\usage{
convert_vector_to_params(vector, oldparams, varnames = NULL)
}
\arguments{
\item{vector}{params}

\item{oldparams}{Only needs to contain "Ns" entry}

\item{varnames}{Optional varnames}
}
\description{
Convert vector to parameter list
}
