% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_longitudinal_cdm.R
\name{fit_longitudinal_cdm}
\alias{fit_longitudinal_cdm}
\title{MCMC Sampler}
\usage{
fit_longitudinal_cdm(
  Ys,
  Xs,
  Qs,
  M,
  priors = list(beta_prior = 1, gamma_prior = 1),
  initparams = NULL
)
}
\arguments{
\item{Ys}{List of response data}

\item{Xs}{Design matrices for gamma}

\item{Qs}{List of Q matrices}

\item{M}{Number of MCMC iterations}

\item{priors}{priors}

\item{initparams}{Initial parameters (optional)}
}
\description{
Add detail
}
