% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygamma_functions.R
\name{gamma_to_transprobs}
\alias{gamma_to_transprobs}
\title{Convert gamma to transition probabilities}
\usage{
gamma_to_transprobs(gamma_list, Xs)
}
\arguments{
\item{gamma_list}{List of gamma parameters}

\item{Xs}{list of design matrices}
}
\description{
Add detail
}
