% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_init_params.R
\name{gen_initial_values_longitudinal}
\alias{gen_initial_values_longitudinal}
\title{Generate initial values from longitudinal cdm model}
\usage{
gen_initial_values_longitudinal(
  Nrespondents,
  Q,
  Ntime = 2,
  Nrespcov = 0,
  Ngroupcov = 0,
  Ngroup = 2,
  seed = NULL
)
}
\arguments{
\item{Nrespondents}{Input values}

\item{Q}{Q-matrix}

\item{Ntime}{Number of time points}

\item{Nrespcov}{Number of respondent-level covariates}

\item{Ngroupcov}{Number of group-level covariates}

\item{Ngroup}{Number of groups}

\item{seed}{Optional seed}
}
\value{
Returns randomly-generated parameter values
}
\description{
Generate initial values from longitudinal cdm model
}
