% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qinfo.R
\name{gen_q_info}
\alias{gen_q_info}
\title{Extracts information from Q-matrix to be used in fitting the model}
\usage{
gen_q_info(Q)
}
\arguments{
\item{Q}{Q-matrix}
}
\value{
Returns a list containing interaction_in_profile,
interaction_qids, interaction_list, skill_in_profile, profile_list,
and q_profiles
}
\description{
Extracts information from Q-matrix to be used in fitting the model
}
