% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_funs.R
\name{gen_u}
\alias{gen_u}
\title{Generate level 2 residual values from beta and gamma parameters}
\usage{
gen_u(params, Xdata)
}
\arguments{
\item{params}{Input values}

\item{Xdata}{Data list}
}
\value{
List containing "uforward" and "ubackward" matrices
}
\description{
Generate level 2 residual values from beta and gamma parameters
}
