% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_functions.R
\name{generate_data}
\alias{generate_data}
\title{Generate data from betas and alphas}
\usage{
generate_data(Nrespondents, Nq_list, beta_mat, alpha_mat)
}
\arguments{
\item{Nrespondents}{Nrespondents}

\item{Nq_list}{Nq_list}

\item{beta_mat}{beta_mat}

\item{alpha_mat}{alpha_mat}
}
\description{
Add detail
}
