% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_functions.R
\name{generate_params}
\alias{generate_params}
\title{Generate random params}
\usage{
generate_params(Nrespondents, Qs, Xs)
}
\arguments{
\item{Nrespondents}{Nrespondents}

\item{Qs}{Qs}

\item{Xs}{Xs}
}
\description{
Add detail
}
