% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{likelihood_master}
\alias{likelihood_master}
\title{Likelihood function for generic longitudinal multi-level cdm model}
\usage{
likelihood_master(params, Xdata, q_info)
}
\arguments{
\item{params}{parameter list}

\item{Xdata}{data list}

\item{q_info}{Result from applying gen_qinfo with Q matrix}
}
\value{
Returns likelihood value
}
\description{
Likelihood function for generic longitudinal multi-level cdm model
}
