% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler.R
\name{mcmc_sampler_main}
\alias{mcmc_sampler_main}
\title{Run MCMC sampler for longitudinal cdm model}
\usage{
mcmc_sampler_main(Xdata, initparams, M)
}
\arguments{
\item{Xdata}{Data list}

\item{initparams}{Initial Parameters}

\item{Q}{Q-matrix}
}
\value{
Returns dataframe of MCMC samples
}
\description{
Run MCMC sampler for longitudinal cdm model
}
