% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_functions.R
\name{sample_alpha_from_gamma}
\alias{sample_alpha_from_gamma}
\title{Sample alpha from gamma}
\usage{
sample_alpha_from_gamma(Nrespondents, Ntime, gamma_list, Xs)
}
\arguments{
\item{Nrespondents}{Nrespondents}

\item{Ntime}{Ntime}

\item{gamma_list}{gamma_list}

\item{Xs}{Xs}
}
\description{
Add detail
}
