% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygamma_functions.R
\name{sample_alpha_mat}
\alias{sample_alpha_mat}
\title{Sample alpha matrix}
\usage{
sample_alpha_mat(alpha_mat, theta, trans_probs, qt_map, profiles, Ys)
}
\arguments{
\item{alpha_mat}{Alpha matrix}

\item{theta}{Response probabilities}

\item{trans_probs}{Transition probabilities}

\item{qt_map}{Map between questions and time}

\item{profiles}{List of profiles}

\item{Ys}{Data}
}
\description{
Add detail
}
