% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygamma_functions.R
\name{sample_gamma}
\alias{sample_gamma}
\title{Sample gamma}
\usage{
sample_gamma(tstar, gamma_list, trans_mat, Xs, priorsd_gamma, retmean = F)
}
\arguments{
\item{tstar}{tstar}

\item{gamma_list}{gamma_list}

\item{trans_mat}{trans_mat}

\item{Xs}{Xs}

\item{priorsd_gamma}{priorsd_gamma}

\item{retmean}{retmean}
}
\description{
Add detail
}
