% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygamma_functions.R
\name{sample_ystar}
\alias{sample_ystar}
\title{Sample polya-gamma auxillary variables}
\usage{
sample_ystar(psi, nct, qt_map)
}
\arguments{
\item{psi}{See Balamuta paper}

\item{nct}{See Balamuta paper}

\item{qt_map}{Map from q to t}
}
\description{
Add detail
}
