% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_cdm_data}
\alias{simulate_cdm_data}
\title{Simulate CDM Xdata}
\usage{
simulate_cdm_data(
  Nrespondents = 20,
  Nquestions = 22,
  Nskill = 3,
  Ntime = 1,
  Ngroup = 1,
  respondent_designmat = Matrix(rep(1, Nrespondents), Nrespondents, 1),
  group_designmat = Matrix(rep(1, Ngroup), Ngroup, 1),
  group_assignments = NULL,
  seed = NULL,
  true_params = NULL,
  true_skill_transprob = NULL,
  Q = NULL,
  true_profiles = NULL
)
}
\arguments{
\item{Nskill}{Number of skills to simulate}

\item{seed}{Fix seed for reproducibility}
}
\value{
Returns simulated X and Q-matrix
}
\description{
Simulate CDM Xdata
}
