% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygamma_functions.R
\name{transitions_to_alpha}
\alias{transitions_to_alpha}
\title{Convert transition matrix to alpha matrix}
\usage{
transitions_to_alpha(trans_mat, transitions)
}
\arguments{
\item{trans_mat}{transition matrix}

\item{transitions}{list of transitions}
}
\description{
Add detail
}
